# backend/storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: KernelLearn API
  version: 1.0.0
  description: |
    API del proyecto KernelLearn - Sprint 1
servers:
  - url: http://100.107.236.28/api/v1
    description: KernelLearn Production Server
  - url: http://localhost:8000/api/v1
    description: Local Development
paths:
...
  /courses:
    get:
      tags:
        - Cursos
      summary: Listar cursos publicados
      operationId: getCourses
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Cursos obtenidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      type: object
                  meta:
                    type: object
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Token de Sanctum (sin前缀 Bearer)